chore: Switch from yarn to npm for organizational consistency
In so doing, rather than trying to port the yarn.lock to package-lock.json,
it updated all the dependencies subject to their constraints in package.json.
I changed any fixed versions to `^` for flexibility and to get the newest
patch versions.

This introduced an error in SetExpiration.ts, related to popper.js, fixed
via mocking that library.