refactor: sign group "timestamp" is a string
Object keys must be strings in JavaScript, but this value which was used
as an object key was sometimes assigned a `number` type. Having it
consistently be a string avoids the need to convert back and forth, and
naming it `key` / `groupKey` makes it clearer that this is an arbitrary
"unique ID" — when the client generates a new key it happens to use the
current Unix timestamp, but keys are never intended to be _interpreted_
as timestamps.